Fix 9 security, correctness, and reliability issues
- CORS origins now env-configurable (COLOR_ANALYSIS_CORS_ALLOWED_ORIGINS)
- Admin token uses hmac.compare_digest; warns at startup when default is set
- Worker skips photos that fail R2 fetch instead of feeding empty bytes
- Worker error handler rolls back before setting failed status
- Redis key deletion uses SCAN instead of blocking KEYS
- R2 session listing paginates via ContinuationToken
- Frontend polling adds 120-attempt ceiling and exponential backoff
- analysis_sessions.reliability changed from Numeric(4,3) to Float with migration

+++ apps/api/src/color_analysis/db/models/analysis_session.py
@@ -1,7 +1,7 @@
 import uuid
 from datetime import datetime, timedelta, timezone
 
-from sqlalchemy import DateTime, ForeignKey, Numeric, String, func
+from sqlalchemy import DateTime, Float, ForeignKey, String, func
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -17,7 +17,7 @@ class AnalysisSession(Base):
     )
     status: Mapped[str] = mapped_column(String(32), default="pending")
     result_state: Mapped[str | None] = mapped_column(String(64), nullable=True)
-    reliability: Mapped[float | None] = mapped_column(Numeric(4, 3), nullable=True)
+    reliability: Mapped[float | None] = mapped_column(Float, nullable=True)
     reliability_bucket: Mapped[str | None] = mapped_column(String(16), nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
     expires_at: Mapped[datetime] = mapped_column(
